+++ src/Common/filesystemHelpers.cpp
@@ -130,28 +130,6 @@ bool pathStartsWith(const String & path, const String & prefix_path)
     return pathStartsWith(filesystem_path, filesystem_prefix_path);
 }
 
-String SQLiteDatabaseValidatePath(const String & path, const String & user_files_path)
-{
-    String canonical_user_files_path = fs::canonical(user_files_path);
-
-    String canonical_path;
-    std::error_code err;
-
-    if (fs::path(path).is_relative())
-        canonical_path = fs::canonical(fs::path(user_files_path) / path, err);
-    else
-        canonical_path = fs::canonical(path, err);
-
-    if (err)
-        throw Exception(ErrorCodes::PATH_ACCESS_DENIED, "SQLite database path '{}' is invalid. Error: {}", path, err.message());
-
-    if (!canonical_path.starts_with(canonical_user_files_path))
-        throw Exception(ErrorCodes::PATH_ACCESS_DENIED,
-                        "SQLite database file path '{}' must be inside 'user_files' directory", path);
-
-    return canonical_path;
-}
-
 }
